+++ src/main/scala/longevity/ddd/subdomain/SupportingSubdomain.scala
@@ -1,9 +1,10 @@
-package longevity.subdomain
+package longevity.ddd.subdomain
 
+import longevity.subdomain.Subdomain
 import longevity.subdomain.embeddable.ETypePool
 import longevity.subdomain.ptype.PTypePool
 
-/** a core domain. functionally equivalent to a [[Subdomain]]
+/** a supporting subdomain. functionally equivalent to a [[Subdomain]].
  *
  * @param name the name of the subdomain
  * @param pTypePool a complete set of the persistent types in the subdomain.
@@ -11,16 +12,18 @@ import longevity.subdomain.ptype.PTypePool
  * @param entityTypePool a complete set of the entity types within the
  * subdomain. defaults to empty
  */
-class CoreDomain(
+class SupportingSubdomain(
   name: String,
   pTypePool: PTypePool = PTypePool.empty,
   entityTypePool: ETypePool = ETypePool.empty)
 extends Subdomain(name, pTypePool, entityTypePool)
 
-/** provides a factory method for constructing [[CoreDomain core domains]] */
-object CoreDomain {
+/** provides a factory method for constructing [[SupportingSubdomain supporting
+ * subdomains]]
+ */
+object SupportingSubdomain {
 
-  /** constructs a core domain. functionally equivalent to a [[Subdomain]].
+  /** constructs a supporting subdomain. functionally equivalent to a [Subdomain].
    *
    * @param name the name of the core domain
    * @param pTypePool a complete set of the persistent types in the subdomain.
@@ -34,7 +37,7 @@ object CoreDomain {
     name: String,
     pTypePool: PTypePool = PTypePool.empty,
     entityTypePool: ETypePool = ETypePool.empty)
-  : CoreDomain = 
-    new CoreDomain(name, pTypePool, entityTypePool)
+  : SupportingSubdomain = 
+    new SupportingSubdomain(name, pTypePool, entityTypePool)
 
 }